d_a_npc_mn: getPosNo u32 count[8]+int loop (66.7->86.4), getRand/getPrmNpcNo int return (84.6->99.6 / 73.8->100), roomId int

+++ include/d/actor/d_a_npc_mn.h
@@ -120,7 +120,7 @@ public:
     void chkMsg();
     void setMessage(unsigned long);
     void setAnmFromMsgTag();
-    s8 getPrmNpcNo();
+    int getPrmNpcNo();
     u8 getPrmRailID();
     u8 getPrmSwitchBit();
     u8 getPrmSwitchBit2();
@@ -133,7 +133,7 @@ public:
     void setAnm(unsigned char, int, float);
     s32 setAnmTbl(sMnAnmDat*);
     s32 XyCheckCB(int);
-    u8 getRand(int);
+    int getRand(int);
     void setCollision(dCcD_Cyl*, cXyz, float, float);
     s32 chkEndEvent();
     u8 chkPosNo();
